fix: move interface Datasource to types folder
