+++ src/types/Datasource.ts
@@ -1,6 +1,6 @@
 import type * as RDF from 'rdf-js';
 import { IConfig } from '../config';
-import Member from '../types/Member';
+import Member from './Member';
 
 interface Datasource {
 	getData(config: IConfig): Promise<Member[]>;
